pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.115.12"}
python-jose = "^3.4.0"
bcrypt = "^4.3.0"
sqlalchemy = "^2.0.40"
asyncpg = "^0.30.0"
alembic = "^1.15.2"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
